Refactored to a more julia-styled version
